+++ src/index.ts
@@ -1,5 +1,4 @@
 import Router from './router/router';
-import Config from './config';
 import WebComponent from './component/web_component';
 
 export * as i18n from './intl/i18n';
@@ -7,6 +6,5 @@ export { translate } from './intl/i18n';
 
 export { html } from './template/template';
 export { createState } from './state/state';
-export { registerCss } from './css';
 export { defineWebComponent } from './component/web_component';
-export { Router, Config, WebComponent };
+export { Router, WebComponent };